TST: isolate flight unit tests
